Added logick blog-cli, blog-was and ref blog-server

+++ blog-server/src/data/post_repository.rs
@@ -1,4 +1,3 @@
-use actix_web::web::post;
 use async_trait::async_trait;
 use sqlx::{query_as, PgPool};
 use crate::domain::error::BlogError;
@@ -23,7 +22,7 @@ impl PostRepository for PostgresPostRepo {
     async fn create_post(&self, author_id: UserId, post: CreatePost) -> Result<ResponsePost, BlogError> {
         let new_post = query_as!(
             ResponsePost,
-            "INSERT INTO posts (title, content, author_id) VALUES ($1, $2, $3) RETURNING title, content",
+            "INSERT INTO posts (title, content, author_id) VALUES ($1, $2, $3) RETURNING id, title, content, author_id",
             post.title,
             post.content,
             author_id.as_uuid(),
@@ -37,7 +36,7 @@ impl PostRepository for PostgresPostRepo {
     async fn get_post_by_id(&self, id: PostId) -> Result<ResponsePost, BlogError> {
         let post = query_as!(
             ResponsePost,
-            "SELECT title, content FROM posts WHERE id = $1",
+            "SELECT id, title, content, author_id FROM posts WHERE id = $1",
             id.as_uuid()
         )
             .fetch_one(&self.pool)
@@ -52,7 +51,7 @@ impl PostRepository for PostgresPostRepo {
 
         let posts = query_as!(
             ResponsePost,
-            "SELECT title, content FROM posts ORDER BY id DESC LIMIT $1 OFFSET $2",
+            "SELECT id, title, content, author_id FROM posts ORDER BY id DESC LIMIT $1 OFFSET $2",
             limit,
             offset
         )
@@ -65,7 +64,7 @@ impl PostRepository for PostgresPostRepo {
     async fn update_post_by_id(&self, author_id: UserId, post_id: PostId, payload: CreatePost) -> Result<ResponsePost, BlogError> {
         let post = query_as!(
             ResponsePost,
-            "UPDATE posts SET title = $1, content = $2 WHERE id = $3 AND author_id = $4 RETURNING title, content",
+            "UPDATE posts SET title = $1, content = $2 WHERE id = $3 AND author_id = $4 RETURNING id, title, content, author_id",
             payload.title,
             payload.content,
             post_id.as_uuid(),
@@ -77,9 +76,9 @@ impl PostRepository for PostgresPostRepo {
     }
 
     async fn delete_post_by_id(&self, author_id: UserId, post_id: PostId) -> Result<(), BlogError> {
-        let post = query_as!(
+        let _post = query_as!(
             ResponsePost,
-            "DELETE FROM posts WHERE id = $1 AND author_id = $2",
+            "DELETE FROM posts WHERE id = $1 AND author_id = $2 RETURNING id, title, content, author_id",
             post_id.as_uuid(),
             author_id.as_uuid(),
         )
